# Connect operations client lazily