--- a/lib/hawkular/operations/operations_api.rb
+++ b/lib/hawkular/operations/operations_api.rb
@@ -1,3 +1,6 @@
+require 'timeout'
+require 'monitor'
+
 require 'hawkular/base_client'
 require 'websocket-client-simple'
 require 'json'
@@ -30,6 +33,7 @@ module Hawkular::Operations
   # Client class to interact with the agent via websockets
   class Client < Hawkular::BaseClient
     include WebSocket::Client
+    include MonitorMixin
 
     attr_accessor :ws, :session_id, :logger
 
@@ -79,22 +83,31 @@ module Hawkular::Operations
 
       super(args[:host], args[:credentials], args[:options])
 
-      url = "ws#{args[:use_secure_connection] ? 's' : ''}://#{args[:host]}/hawkular/command-gateway/ui/ws"
+      @logger = Hawkular::Logger.new
 
-      creds = args[:credentials]
-      base64_creds = ["#{creds[:username]}:#{creds[:password]}"].pack('m').delete("\r\n")
+      @url = "ws#{args[:use_secure_connection] ? 's' : ''}://#{args[:host]}/hawkular/command-gateway/ui/ws"
+      @credentials = args[:credentials]
+      @tenant = args[:options][:tenant]
+      @wait_time = args[:wait_time]
+    end
+
+    def base64_credentials
+      ["#{@credentials[:username]}:#{@credentials[:password]}"].pack('m').delete("\r\n")
+    end
+
+    def connect
+      return if synchronize { @connecting ||= false }
+      @connecting = true
 
       ws_options = {
         headers:  {
-          'Authorization' => 'Basic ' + base64_creds,
-          'Hawkular-Tenant' => args[:options][:tenant],
+          'Authorization' => 'Basic ' + base64_credentials,
+          'Hawkular-Tenant' => @tenant,
           'Accept' => 'application/json'
         }
       }
 
-      @logger = Hawkular::Logger.new
-
-      @ws = Simple.connect url, ws_options do |client|
+      @ws = Simple.connect @url, ws_options do |client|
         client.on(:message, once: true) do |msg|
           parsed_message = msg.data.to_msg_hash
 
@@ -106,13 +119,11 @@ module Hawkular::Operations
           end
         end
       end
-
-      sleep args[:wait_time]
     end
 
     # Closes the WebSocket connection
     def close_connection!
-      @ws.close
+      @ws && @ws.close
     end
 
     # Invokes a generic operation on the WildFly agent
@@ -322,6 +333,9 @@ module Hawkular::Operations
     private
 
     def invoke_operation_helper(operation_payload, operation_name = nil, binary_content = nil, &callback)
+      connect unless @connecting
+      sleep @wait_time
+
       # fallback to generic 'ExecuteOperation' if nothing is specified
       operation_name ||= 'ExecuteOperation'
       add_credentials! operation_payload
@@ -332,10 +346,11 @@ module Hawkular::Operations
       payload = "#{operation_name}Request=#{operation_payload.to_json}"
       payload += binary_content unless binary_content.nil?
       @ws.send payload, type: binary_content.nil? ? :text : :binary
+    rescue => e
+      callback.perform(:failure, "#{e.class} - #{e.message}")
     end
 
     def check_pre_conditions(hash = {}, params = [], &callback)
-      fail 'Handshake with server has not been done.' unless @ws.open?
       fail 'Hash cannot be nil.' if hash.nil?
       fail 'callback must have the perform method defined. include Hawkular::Operations' unless
           callback.nil? || callback.respond_to?('perform')