--- a/lib/hawkular/operations/operations_api.rb
+++ b/lib/hawkular/operations/operations_api.rb
@@ -18,7 +18,6 @@ class Proc
       define_method("#{method_name}?") { true }
 
       # method_missing is here because we are not forcing the client to provide both success and error callbacks
-      # rubocop:disable Lint/NestedMethodDefinition
       # https://github.com/bbatsov/rubocop/issues/2704
       def method_missing(_method_name, *_args, &_block)
         PerformMethodMissing
@@ -79,7 +78,7 @@ module Hawkular::Operations
         args[:use_secure_connection] = %w(https wss).include?(uri.scheme) ? true : false
       end
 
-      fail 'no parameter ":host" or ":entrypoint" given' if args[:host].nil?
+      raise 'no parameter ":host" or ":entrypoint" given' if args[:host].nil?
 
       super(args[:host], args[:credentials], args[:options])
 
@@ -96,7 +95,8 @@ module Hawkular::Operations
     end
 
     def connect
-      return if synchronize { @connecting ||= false }
+      return if @connecting || (@ws && @ws.open?)
+
       @connecting = true
 
       ws_options = {
@@ -119,6 +119,10 @@ module Hawkular::Operations
           end
         end
       end
+
+      Timeout.timeout(@wait_time) { sleep 0.1 until @ws.open? }
+    ensure
+      @connecting = false
     end
 
     # Closes the WebSocket connection
@@ -147,7 +151,7 @@ module Hawkular::Operations
     # RemoveDatasource (and not RemoveDatasourceRequest)
     # @param callback [Block] callback that is run after the operation is done
     def invoke_specific_operation(operation_payload, operation_name, &callback)
-      fail 'Operation must be specified' if operation_name.nil?
+      raise 'Operation must be specified' if operation_name.nil?
       required = [:resourcePath]
       check_pre_conditions operation_payload, required, &callback
 
@@ -308,7 +312,7 @@ module Hawkular::Operations
     # @param [String] resource_path canonical path of the WildFly server
     # @param callback [Block] callback that is run after the operation is done
     def export_jdr(resource_path, &callback)
-      fail 'resource_path must be specified' if resource_path.nil?
+      raise 'resource_path must be specified' if resource_path.nil?
       check_pre_conditions(&callback)
 
       invoke_specific_operation({ resourcePath: resource_path }, 'ExportJdr', &callback)
@@ -333,8 +337,7 @@ module Hawkular::Operations
     private
 
     def invoke_operation_helper(operation_payload, operation_name = nil, binary_content = nil, &callback)
-      connect unless @connecting
-      sleep @wait_time
+      synchronize { connect }
 
       # fallback to generic 'ExecuteOperation' if nothing is specified
       operation_name ||= 'ExecuteOperation'
@@ -351,15 +354,15 @@ module Hawkular::Operations
     end
 
     def check_pre_conditions(hash = {}, params = [], &callback)
-      fail 'Hash cannot be nil.' if hash.nil?
-      fail 'callback must have the perform method defined. include Hawkular::Operations' unless
+      raise 'Hash cannot be nil.' if hash.nil?
+      raise 'callback must have the perform method defined. include Hawkular::Operations' unless
           callback.nil? || callback.respond_to?('perform')
       params.each do |property|
         next unless hash[property].nil?
         err_callback = 'You need to specify error callback'
         err_message = "Hash property #{property} must be specified"
-        fail(ArgumentError, err_callback) if callback.nil?
-        fail(ArgumentError, err_callback) if callback.perform(:failure, err_message).equal? Proc::PerformMethodMissing
+        raise(ArgumentError, err_callback) if callback.nil?
+        raise(ArgumentError, err_callback) if callback.perform(:failure, err_message).equal? Proc::PerformMethodMissing
       end
     end