… spec/vcr_cassettes/Operation/NonSecure/Websocket_connection/connects_correctly.json
… spec/vcr_cassettes/Operation/Secure/Websocket_connection/connects_correctly.json