This pull request fixes a 500 page that happens when you try to do a power
operation when the middleware server is offline, by making the connection be
created lazily when an operation is called.

The idea behind this is to allow operations code to catch errors when connecting to the server, because while errors from the operations can be catched.

As an example, here's this piece of code:

```ruby
connection.operations(true).undeploy(deployment_data) do |on|
  on.success do |data|
    _log.debug "Success on websocket-operation #{data}"
    emit_middleware_notification(:mw_op_success, 'Undeploy', deployment_name, ems_ref, MiddlewareDeployment)
  end
  on.failure do |error|
    _log.error 'error callback was called, reason: ' + error.to_s
    emit_middleware_notification(:mw_op_failure, 'Undeploy', deployment_name, ems_ref, MiddlewareDeployment)
  end
end
```

If the error happens on `#undeploy`, then the failure clause will be called, but if the error happens on `#operations(true)`, which spawns a new connection, it will be raised as normal.

As a bonus, this PR also removes the sleep code from the connection handler,
which makes both the tests and real world usage much, much, much faster.

Before:

        hawkular-client-ruby git:(fix-operations-500-error) $ time RUBYOPT="-Ilib -W0 -r./bundle/bundler/setup.rb" rspec spec/integration/operations_spec.rb
        [Coveralls] Set up the SimpleCov formatter.
        [Coveralls] Using SimpleCov's default settings.
        Run options: exclude {:skip=>true}
        ...................

        Finished in 32.57 seconds (files took 0.69792 seconds to load)
        19 examples, 0 failures

After:

        hawkular-client-ruby git:(fix-operations-500-error) ✗ $ time RUBYOPT="-Ilib -W0 -r./bundle/bundler/setup.rb" rspec spec/integration/operations_spec.rb
        [Coveralls] Set up the SimpleCov formatter.
        [Coveralls] Using SimpleCov's default settings.
        Run options: exclude {:skip=>true}
        ...................

        Finished in 3.34 seconds (files took 0.74947 seconds to load)
        19 examples, 0 failures

Bugzilla Ticket: https://bugzilla.redhat.com/show_bug.cgi?id=1444940