@cfcosta 2017-06-14
@israel-hdez @abonas did some changes, can you take a look at it again?