@pilhuhn 2017-06-19
I think that makes sense. I recall that we battled with getting the error callback right in the past for cases when credentials are wrong and similar.