@israel-hdez 2017-06-20
@cfcosta I have no more comments. Ready to merge?